Tests(AuthModule): Change auth module's mocked data name
Description:
To enable the creation of consistent mock data with varying values
across different tests, I've added the module name (to which the mocked
data belongs) as an extension to the mocked data. This change ensures
the ability to separate and manage mocked data for each module without
affecting others.